GenDoc

Documentation for GenDoc.

GenDoc.generate_front_matter — Method

generate_front_matter(;kwargs...)

Generates YAML front matter based on given vars.

GenDoc.md — Method

md(df::DataFrame)

Convert DataFrame to String.

GenDoc.rplot — Method

rplot(filename; path_prefix, uri_prefix, kwargs...)::String

Wrapper around the ggsave function from Tidyverse. Returns a string pointing to the image generated by ggsave. The image is created at path_prefix / filename. To avoid constantly passing path_prefix and url_prefix, consider creating a partial function.

GenDoc.@withcb — Macro

@withcb ex

Returns the output of the evaluated expression eval(ex) after the expression ex inside a Markdown code block. Returns the expression ex inside a Markdown code block and the output of eval(ex).

Similar to default behaviour of R Markdown code blocks.

MacroTools.rmlines should help.