Methods

Public

GenDoc.generate_front_matter — Method

generate_front_matter(; kwargs...)

Generates YAML front matter based on given keyword arguments kwargs.

GenDoc.hspace — Function

hspace(h::Number)::String
hspace(h="1cm")::String

Horizontal space as HTML. The unit of h::Number defaults to cm.

GenDoc.is_gitlab_ci — Method

is_gitlab_ci()::Bool

Return whether this project is being executed on GitLab's CI.

GenDoc.md — Method

md(df::DataFrame)::String

Convert DataFrame df to Markdown syntax.

GenDoc.pretty — Method

pretty(df::DataFrame)::String

Make header of df prettier, by capitalising the first letter and removing underscores.

GenDoc.rplot — Method

rplot(filename; path_prefix=nothing, uri_prefix=nothing, kwargs...)::String

Wrapper around the ggsave function from Tidyverse. Keyword arguments kwargs are passed to ggsave. Returns a string pointing to the image generated by ggsave. The image is created at path_prefix / filename. To avoid constantly passing path_prefix and url_prefix, consider creating a partial function.

GenDoc.@withcb — Macro

@withcb ex

Return the expression ex inside a Markdown code block and the output of eval(ex).

Similar to default behaviour of R Markdown code blocks.

Private

GenDoc.project_root — Method

project_root()::String

Return the root directory of the project.